api/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pydantic>=2.6.4
pydantic-settings>=2.2.1
email-validator>=2.2.0
python-dotenv>=1.0.1
starlette>=0.36.0
python-multipart>=0.0.9
bcrypt==4.1.3
passlib>=1.7.4
python-jose>=3.3.0
cryptography>=42.0.8
tzdata>=2024.2
sqlalchemy>=2.0.29
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
mangum>=0.17.0
openpyxl>=3.1.2
httpx>=0.27.0
aiosqlite>=0.20.0
greenlet>=3.0.0

